Change the default 'main' file for catalyst
when the repo is initially cloned or installed, ./catalyst.css doesn't exist yet, so the 'main' entry in package.json is invalid.
The user of the library first needs to run 'make build' or 'make dist' to generate ./catalyst.css. Howver, there is a
(presumably stable) version of catalyst bundled with the repo by default, dist/catalyst.css. This commit changes the
'main' entry to point to that default version